+++ jgclark.Dashboard/CHANGELOG.md
@@ -16,6 +16,7 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
 - Fix edge case on display of URLs with embedded ~ characters
 - Fix tags being missed when they have a different case (only applies when the new cache isn't being used)
 - Smarter refresh of display after moving an item between notes
+- the Priority and Overdue sections now applies the 'Apply to sections under Calendar note headings' sub-setting of 'Ignore items in notes with phrase(s)' where set
 - will now trigger a tagCache rebuild (if scheduled) after a refresh, as well as after first load
 - now rebuilds tagCache after adding a new tag to a perspective before its saved
 - fixes when changing perspective settings (DBW)